Deleted Duplicate game_board
Fixed the issue of there being multiple game_board files. It is now in the map folder.
