+++ game/common/map/game_board.py
@@ -1,6 +1,8 @@
 import random
 from typing import Self
 from game.utils.vector import Vector
+from game.common.stations.occupiable_station import Occupiable_Station
+from game.common.stations.station import Station
 from game.common.avatar import Avatar
 from game.common.game_object import GameObject
 from game.common.map.tile import Tile
@@ -17,18 +19,29 @@ class GameBoard(GameObject):
         For example, a Vector object with an 'x' of 5 and a 'y' of 7 will create a board 5 tiles wide and
         7 tiles long
 
+        Example:
+            _ _ _ _ _  y = 0
+            |       |
+            |       |
+            |       |
+            |       |
+            |       |
+            |       |
+            _ _ _ _ _  y = 6
+
     -----------------------------------------------------------------------------------------------------------
 
     locations:
         This is the bulkiest part of the generation. The locations field is a dictionary with a key
-        being a list of Vectors, and the value being a list of GameObjects. This is used to assign the
-        given GameObjects the given coordinates via the Vectors. This is done in two ways:
+        being a tuple of Vectors, and the value being a list of GameObjects (the key is a tuple to prevent issues
+        with the generation system). This is used to assign the given GameObjects the given coordinates
+        via the Vectors. This is done in two ways:
 
             Statically:
                 If you want a GameObject to be at a specific coordinate, ensure that the key-value pair is
                 ONE Vector and ONE GameObject.
                 An example of this would be the following:
-                    locations = {[vector_2_4] : [station_0]}
+                    locations = {(vector_2_4) : [station_0]}
 
                 In this example, vector_2_4 contains the coordinates (2, 4). (Note that this naming convention
                 isn't necessary, but was used to help with the concept). Furthermore, station_0 is the
@@ -36,29 +49,31 @@ class GameBoard(GameObject):
 
             Dynamically:
                 If you want to assign multiple GameObjects to different coordinates, use a key-value
-                pair of any length. However, the assignments will be random.
-                An example of this would be the following:
+                pair of any length. NOTE: The length of the tuple and list MUST be equal, otherwise it will not work.
+                In this case, the assignments will be random. An example of this would be the following:
                     locations =
                         {
-                            [vector_0_0, vector_1_1, vector_2_2] : [station_0, station_1, station_2]
+                            (vector_0_0, vector_1_1, vector_2_2) : [station_0, station_1, station_2]
                         }
 
+                (Note that the tuple and list both have a length of 3).
                 When this is passed in, the three different vectors containing coordinates (0, 0), (1, 1), or
                 (2, 2) will be randomly assigned station_0, station_1, or station_2.
 
                 Therefore:
-                If station_0 is at (1, 1),
+                If station_0 is randomly assigned at (1, 1),
                 station_1 could be at (2, 2),
                 then station_2 will be at (0, 0).
+                This is just one case of what could happen.
 
         Lastly, another example will be shown to explain that you can combine both static and
         dynamic assignments in the same dictionary:
 
             locations =
                 {
-                    [vector_0_0] : [station_0],
-                    [vector_0_1] : [station_1],
-                    [vector_1_1, vector_1_2, vector_1_3] : [station_2, station_3, station_4]
+                    (vector_0_0) : [station_0],
+                    (vector_0_1) : [station_1],
+                    (vector_1_1, vector_1_2, vector_1_3) : [station_2, station_3, station_4]
                 }
 
         In this example, station_0 will be at vector_0_0 without interference. The same applies to
@@ -86,62 +101,90 @@ class GameBoard(GameObject):
     """
 
     def __init__(self, seed: int | None = None, map_size: Vector = Vector(),
-                 locations: dict[[Vector]:[GameObject]] | None = None, walled: bool = False):
+                 locations: dict[tuple[Vector]:list[Tile]] | None = None, walled: bool = False):
 
         super().__init__()
-        self.seed = seed
-        random.seed(seed)
-        self.object_type: ObjectType = ObjectType.GAMEBOARD
-        self.event_active = None
-        self.map_size: Vector = map_size
-        self.locations: dict = locations
-        self.walled: bool = walled
-
         # game_map is initially going to be None. Since generation is slow, call generate_map() as needed
         self.game_map: list[list[Tile]] | None = None
+        self.seed: int | None = seed
+        random.seed(seed)
+        self.object_type: ObjectType = ObjectType.GAMEBOARD
+        self.event_active: int | None = None
+        self.map_size: Vector = map_size
+        # when passing Vectors as a tuple, end the tuple of Vectors with a comma so it is recognized as a tuple
+        self.locations: dict | None = locations
+        self.walled: bool = walled
 
     @property
     def seed(self) -> int:
         return self.__seed
 
     @seed.setter
-    def seed(self, seed: int | None):
+    def seed(self, seed: int | None) -> None:
+        if self.game_map is not None:
+            raise RuntimeError(f'{self.__class__.__name__} variables cannot be changed once generate_map is run.')
         if seed is not None and not isinstance(seed, int):
-            raise ValueError("Seed must be an integer.")
+            raise ValueError(f'{self.__class__.__name__}.seed must be an integer.')
         self.__seed = seed
+
+    @property
+    def game_map(self) -> list[list[Tile]] | None:
+        return self.__game_map
+
+    @game_map.setter
+    def game_map(self, game_map: list[list[Tile]]) -> None:
+        if game_map is not None and (not isinstance(game_map, list) or
+                                     any(map(lambda l: not isinstance(l, list), game_map)) or
+                                     any([any(map(lambda g: not isinstance(g, Tile), tile_list))
+                                          for tile_list in game_map])):
+            raise ValueError(f'{self.__class__.__name__}.game_map must be a list[list[Tile]].')
+        self.__game_map = game_map
 
     @property
     def map_size(self) -> Vector:
         return self.__map_size
 
     @map_size.setter
-    def map_size(self, map_size: Vector):
-        if map_size is None or isinstance(map_size, Vector):
-            self.__map_size = map_size
-        raise ValueError("Map_size must be a Vector.")
+    def map_size(self, map_size: Vector) -> None:
+        if self.game_map is not None:
+            raise RuntimeError(f'{self.__class__.__name__} variables cannot be changed once generate_map is run.')
+        if map_size is None or not isinstance(map_size, Vector):
+            raise ValueError(f'{self.__class__.__name__}.map_size must be a Vector.')
+        self.__map_size = map_size
 
     @property
     def locations(self) -> dict:
         return self.__locations
 
     @locations.setter
-    def locations(self, locations: dict[[Vector]:[GameObject]]):
-        if locations is None or isinstance(locations, dict):
-            self.__locations = locations
-        raise ValueError("Locations must be a dict. The key must be a list of Vector Objects, and the "
-                         "value a list of GameObject.")
+    def locations(self, locations: dict[tuple[Vector]:list[GameObject]] | None) -> None:
+        if self.game_map is not None:
+            raise RuntimeError(f'{self.__class__.__name__} variables cannot be changed once generate_map is run.')
+        if locations is not None and not isinstance(locations, dict):
+            raise ValueError("Locations must be a dict. The key must be a tuple of Vector Objects, and the "
+                             "value a list of GameObject.")
+        if locations is not None:
+            for k, v in locations.items():
+                if len(k) != len(v):
+                    raise ValueError("Cannot set the locations for the game_board. A key has a different "
+                                     "length than its key.")
+
+        self.__locations = locations
 
     @property
     def walled(self) -> bool:
         return self.__walled
 
     @walled.setter
-    def walled(self, walled: bool):
-        if walled is None or isinstance(walled, bool):
-            self.__walled = walled
-        raise ValueError("Walled must be a bool.")
+    def walled(self, walled: bool) -> None:
+        if self.game_map is not None:
+            raise RuntimeError(f'{self.__class__.__name__} variables cannot be changed once generate_map is run.')
+        if walled is None or not isinstance(walled, bool):
+            raise ValueError(f'{self.__class__.__name__}.walled must be a bool.')
 
-    def generate_map(self):
+        self.__walled = walled
+
+    def generate_map(self) -> None:
         # generate map
         self.game_map = [[Tile() for _ in range(self.map_size.x)] for _ in range(self.map_size.y)]
 
@@ -155,43 +198,42 @@ class GameBoard(GameObject):
 
         self.__populate_map()
 
-    def __populate_map(self):
+    def __populate_map(self) -> None:
         for k, v in self.locations.items():
             if len(k) != len(v) or (len(k) == 0 or len(v) == 0):  # Key-Value lengths must be > 0 and equal
                 raise ValueError("A key-value pair from game_board.locations has mismatching lengths. "
                                  "They must be the same length, regardless of size.")
-            j = random.choices(k, k=len(k))
+
+            # random.sample returns a randomized list which is used in __help_populate()
+            j = random.sample(k, k=len(k))
             self.__help_populate(j, v)
 
-    def __help_populate(self, vector_list: list[Vector], v: list[GameObject]):
-        for i in v:
-            temp_vector: Vector = vector_list.pop()
-
+    def __help_populate(self, vector_list: list[Vector], v: list[GameObject]) -> None:
+        for j, i in zip(vector_list, v):
             if isinstance(i, Avatar):  # If the GameObject is an Avatar, assign it the coordinate position
-                i.position = temp_vector
+                i.position = j
 
-            temp_tile: GameObject = self.game_map[temp_vector.y][temp_vector.x]
+            temp_tile: GameObject = self.game_map[j.y][j.x]
 
             while hasattr(temp_tile.occupied_by, 'occupied_by'):
                 temp_tile = temp_tile.occupied_by
 
-            if temp_tile is not None:
+            if temp_tile is None:
                 raise ValueError("Last item on the given tile doesn't have the 'occupied_by' attribute.")
 
             temp_tile.occupied_by = i
 
     def get_objects(self, look_for: ObjectType) -> list[GameObject]:
-        to_return = list()
+        to_return: list[GameObject] = list()
 
         for row in self.game_map:
             for object_in_row in row:
-                temp: GameObject | Tile = object_in_row
+                temp: GameObject = object_in_row
                 self.__get_objects_help(look_for, temp, to_return)
 
         return to_return
 
-    @staticmethod
-    def __get_objects_help(look_for: ObjectType, temp: GameObject | Tile, to_return: list[GameObject]):
+    def __get_objects_help(self, look_for: ObjectType, temp: GameObject | Tile, to_return: list[GameObject]):
         while hasattr(temp, 'occupied_by'):
             if temp.object_type is look_for:
                 to_return.append(temp)
@@ -203,26 +245,48 @@ class GameBoard(GameObject):
             to_return.append(temp)
 
     def to_json(self) -> dict:
-        data = super().to_json()
-        temp = list([list(map(lambda tile: tile.to_json(), y)) for y in self.game_map])
+        data: dict[str, object] = super().to_json()
+        temp: list[list[Tile]] = list(
+            list(map(lambda tile: tile.to_json(), y)) for y in self.game_map) if self.game_map is not None else None
         data["game_map"] = temp
         data["seed"] = self.seed
-        data["map_size"] = self.map_size
-        data["locations"] = self.locations
+        data["map_size"] = self.map_size.to_json()
+        data["location_vectors"] = [[vec.to_json() for vec in k] for k in
+                                    self.locations.keys()] if self.locations is not None else None
+        data["location_objects"] = [[obj.to_json() for obj in v] for v in
+                                    self.locations.values()] if self.locations is not None else None
         data["walled"] = self.walled
         data['event_active'] = self.event_active
         return data
 
-    def generate_event(self, start, end):
+    def generate_event(self, start: int, end: int) -> None:
         self.event_active = random.randint(start, end)
 
-    def from_json(self, data) -> Self:
+    def __from_json_helper(self, data: dict) -> GameObject:
+        match data['object_type']:
+            case ObjectType.WALL:
+                return Wall().from_json(data)
+            case ObjectType.OCCUPIABLE_STATION:
+                return Occupiable_Station().from_json(data)
+            case ObjectType.STATION:
+                return Station().from_json(data)
+            case ObjectType.AVATAR:
+                return Avatar().from_json(data)
+            # If adding more ObjectTypes that can be placed on the game_board, specify here
+            case _:
+                raise ValueError(f'The location (dict) must have a valid key (tuple of vectors) and a valid value ('
+                                 f'list of GameObjects).')
+
+    def from_json(self, data: dict) -> Self:
         super().from_json(data)
         temp = data["game_map"]
-        self.game_map = list([list(map(lambda tile: Tile().from_json(tile), y)) for y in temp])
-        self.seed = data["seed"]
-        self.map_size = data["map_size"]
-        self.locations = data["locations"]
-        self.walled = data["walled"]
-        self.event_active = data['event_active']
+        self.seed: int | None = data["seed"]
+        self.map_size: Vector = Vector().from_json(data["map_size"])
+        self.locations: dict[tuple[Vector]:list[Tile]] = {
+            tuple(map(lambda vec: Vector().from_json(vec), k)): [self.__from_json_helper(obj) for obj in v] for k, v in
+            zip(data["location_vectors"], data["location_objects"])} if data["location_vectors"] is not None else None
+        self.walled: bool = data["walled"]
+        self.event_active: int = data['event_active']
+        self.game_map: list[list[Tile]] = list(
+            list(map(lambda tile: Tile().from_json(tile), y)) for y in temp) if temp is not None else None
         return self
